feat: add Playbill job board crawler with valid URL extraction
Introduces a two-stage Playbill crawler that monitors the Musician category
for trumpet-relevant listings. Fixes URL construction by extracting full
href attributes (including UUID slugs) from raw HTML before stripping, then
passing known URLs to Claude for title/org matching rather than letting it
construct URLs from text.

Also refactors shared HTTP/scraping utilities into scraper.ts and all
Playbill-specific logic into playbill-crawler.ts to keep check-auditions.ts
focused on orchestration.
