+++ check-auditions.ts
@@ -25,10 +25,23 @@
 import * as fs from "fs";
 import * as path from "path";
 import * as https from "https";
-import * as http from "http";
-import * as crypto from "crypto";
 import Anthropic from "@anthropic-ai/sdk";
 import { google } from "googleapis";
+
+import {
+  MIN_CONTENT_LENGTH,
+  fetchPage,
+  fetchWithPuppeteer,
+  stripHtml,
+  scrapeUrl,
+  contentHash,
+} from "./scraper";
+import {
+  PlaybillListing,
+  PlaybillFinding,
+  PlaybillState,
+  processPlaybillUrl,
+} from "./playbill-crawler";
 
 // ─── Types ────────────────────────────────────────────────────────────────────
 
@@ -62,51 +75,14 @@ interface PageState {
 }
 
 /**
- * Persistent record for a single Playbill job listing.
- * Because Playbill is a rolling job board, listings are tracked individually by URL
- * rather than by page hash. `notified` is the canonical "done" flag — it is set only
- * after a successful email send, providing at-least-once delivery semantics if the
- * process crashes between trumpet confirmation and email dispatch.
- * Non-trumpet listings are also marked `notified=true` to prevent rechecking them on
- * every subsequent run once we know they aren't relevant.
- */
-interface PlaybillListing {
-  url: string;
-  title: string;
-  organization: string;
-  firstSeen: string;
-  lastChecked: string;
-  hasTrumpet: boolean;
-  notified: boolean;
-  summary: string | null;
-}
-
-/** Transient shape returned by Claude during index-page extraction. Never persisted directly. */
-interface PlaybillIndexListing {
-  title: string;
-  url: string;
-  organization: string;
-}
-
-/** Carries confirmed trumpet hits into the email layer, decoupled from the full PlaybillListing state. */
-interface PlaybillFinding {
-  listingUrl: string;
-  title: string;
-  organization: string;
-  summary: string | null;
-}
-
-/**
  * The full persisted state written to audition-state.json after each run.
- * The two Playbill fields (`playbillIndexHash`, `playbillListings`) are additive —
- * `loadState()` populates them with safe defaults so existing state files without
- * these keys continue to work without a migration step.
+ * The Playbill fields come from PlaybillState and are additive — `loadState()`
+ * populates them with safe defaults so existing state files without these keys
+ * continue to work without a migration step.
  */
-interface StateFile {
+interface StateFile extends PlaybillState {
   lastRun: string;
   pages: Record<string, PageState>;
-  playbillIndexHash: string | null;
-  playbillListings: Record<string, PlaybillListing>;
 }
 
 interface AuditionAnalysis {
@@ -120,130 +96,6 @@ interface AuditionAnalysis {
 
 const STATE_FILE = path.join(process.cwd(), "audition-state.json");
 const URLS_FILE = path.join(process.cwd(), "urls.json");
-const MIN_CONTENT_LENGTH = 500; // chars threshold to consider fetch "empty"
-
-// ─── HTTP Fetch (no deps) ─────────────────────────────────────────────────────
-
-/**
- * Intentionally uses only Node's built-in http/https modules (no axios, node-fetch, etc.)
- * to keep the dependency surface minimal for a script that runs in CI with no build step.
- */
-function fetchPage(url: string): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const client = url.startsWith("https") ? https : http;
-    const req = client.get(
-      url,
-      {
-        headers: {
-          "User-Agent":
-            "Mozilla/5.0 (compatible; AuditionChecker/1.0; +https://github.com)",
-          Accept: "text/html,application/xhtml+xml",
-        },
-      },
-      (res) => {
-        // Handle redirects
-        if (
-          res.statusCode &&
-          res.statusCode >= 300 &&
-          res.statusCode < 400 &&
-          res.headers.location
-        ) {
-          fetchPage(res.headers.location).then(resolve).catch(reject);
-          return;
-        }
-        let data = "";
-        res.on("data", (chunk) => (data += chunk));
-        res.on("end", () => resolve(data));
-      }
-    );
-    req.on("error", reject);
-    req.setTimeout(15000, () => {
-      req.destroy();
-      reject(new Error("Request timed out"));
-    });
-  });
-}
-
-// ─── Puppeteer fallback ───────────────────────────────────────────────────────
-
-/**
- * JS-rendered pages (SPAs, lazy-loaded content) return near-empty HTML via plain HTTP.
- * Puppeteer is the escape hatch for those cases. The dynamic import keeps this optional:
- * the script starts and runs standard URLs even if puppeteer isn't installed, only
- * failing at the point a JS-rendered page is actually needed.
- */
-async function fetchWithPuppeteer(url: string): Promise<string> {
-  // Dynamically import so the script still runs if puppeteer isn't installed
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let puppeteer: any;
-  try {
-    puppeteer = await import("puppeteer");
-  } catch {
-    throw new Error(
-      "Puppeteer not installed. Run: npm install puppeteer  (or add to package.json)"
-    );
-  }
-
-  const browser = await puppeteer.default.launch({
-    headless: true,
-    args: ["--no-sandbox", "--disable-setuid-sandbox"],
-  });
-  try {
-    const page = await browser.newPage();
-    await page.setUserAgent(
-      "Mozilla/5.0 (compatible; AuditionChecker/1.0; +https://github.com)"
-    );
-    await page.goto(url, { waitUntil: "networkidle2", timeout: 30000 });
-    // Wait a bit more for lazy-loaded content
-    await new Promise((r) => setTimeout(r, 2000));
-    return await page.content();
-  } finally {
-    await browser.close();
-  }
-}
-
-// ─── HTML stripper ────────────────────────────────────────────────────────────
-
-function stripHtml(html: string): string {
-  return html
-    .replace(/<script[\s\S]*?<\/script>/gi, "")
-    .replace(/<style[\s\S]*?<\/style>/gi, "")
-    .replace(/<[^>]+>/g, " ")
-    .replace(/&nbsp;/g, " ")
-    .replace(/&amp;/g, "&")
-    .replace(/&lt;/g, "<")
-    .replace(/&gt;/g, ">")
-    .replace(/\s{2,}/g, " ")
-    .trim();
-}
-
-// ─── Scrape with fallback ─────────────────────────────────────────────────────
-
-/**
- * Chain of Responsibility: plain HTTP → Puppeteer, with a content-length gate as
- * the second fallback trigger. The gate catches sites that return 200 OK with a
- * skeleton HTML shell (common with SPAs) even when the plain fetch technically succeeds.
- * All callers receive plain text — the fetch strategy is an implementation detail.
- */
-async function scrapeUrl(url: string): Promise<string> {
-  console.log(`  Fetching: ${url}`);
-  let html: string;
-  try {
-    html = await fetchPage(url);
-  } catch (err) {
-    console.log(`  Fetch failed (${err}), trying Puppeteer...`);
-    html = await fetchWithPuppeteer(url);
-  }
-
-  if (html.length < MIN_CONTENT_LENGTH) {
-    console.log(
-      `  Content too short (${html.length} chars), falling back to Puppeteer...`
-    );
-    html = await fetchWithPuppeteer(url);
-  }
-
-  return stripHtml(html);
-}
 
 // ─── Claude Analysis ──────────────────────────────────────────────────────────
 
@@ -320,238 +172,6 @@ ${pageText.slice(0, 8000)}`;
   }
 }
 
-// ─── Playbill: extract musician listings from index page ──────────────────────
-
-/**
- * Stage 1 of the two-stage Playbill pipeline. Claude acts as a structured data extractor
- * rather than a classifier here: the goal is to pull (title, url, organization) tuples
- * from unstructured HTML-stripped text. The URL normalization step after parsing is
- * necessary because Claude may return relative paths depending on how Playbill renders
- * its links in the stripped text.
- *
- * Errors are swallowed and return [] by design — a bad Claude response on the index
- * should not halt the run or suppress notifications from other sources.
- */
-async function extractPlaybillListings(
-  client: Anthropic,
-  indexText: string
-): Promise<PlaybillIndexListing[]> {
-  const prompt = `You are helping a professional trumpet player find performing musician job listings.
-
-The page below is pre-filtered to the "Musician" category on Playbill. Extract all listed job postings,
-but EXCLUDE any that are clearly non-performing roles such as: Music Director, Conductor,
-Director of Music, Music Administrator, or similar leadership/administrative titles.
-Also EXCLUDE any unpaid, volunteer, internship, or stipend-only positions. Only include paid jobs.
-
-Return a JSON array only (no prose):
-[{ "title": string, "url": string, "organization": string }]
-
-- url: the absolute URL to the individual listing (prepend https://playbill.com if relative or no host)
-- Return [] if no relevant listings found.
-
-Page content (first 6000 chars):
-${indexText.slice(0, 6000)}`;
-
-  try {
-    const response = await client.messages.create({
-      model: "claude-sonnet-4-20250514",
-      max_tokens: 2000,
-      messages: [{ role: "user", content: prompt }],
-    });
-
-    const raw = response.content
-      .filter((b) => b.type === "text")
-      .map((b) => (b as { type: "text"; text: string }).text)
-      .join("");
-
-    const match = raw.match(/\[[\s\S]*\]/);
-    if (!match) throw new Error("No JSON array found");
-
-    const listings = JSON.parse(match[0]) as PlaybillIndexListing[];
-
-    // Normalize relative URLs and drop malformed entries
-    return listings
-      .filter((l) => l && typeof l.title === "string" && typeof l.url === "string" && typeof l.organization === "string")
-      .map((l) => ({
-        ...l,
-        url: l.url.startsWith("http") ? l.url : `https://playbill.com${l.url.startsWith("/") ? "" : "/"}${l.url}`,
-      }));
-  } catch (err) {
-    console.warn(`  ⚠️  extractPlaybillListings failed: ${err}`);
-    return [];
-  }
-}
-
-// ─── Playbill: check if a listing detail page mentions trumpet ─────────────────
-
-/**
- * Stage 2 of the Playbill pipeline. Claude acts as a binary classifier against a
- * single listing's full description. This is intentionally a separate Claude call from
- * Stage 1 because the index page has only listing metadata (title, org), while
- * instrument requirements are buried in the detail page body. The two-call design
- * avoids fetching every detail page on every run — Stage 1 gates which listings
- * even reach Stage 2 via the `notified` flag in state.
- */
-async function checkListingForTrumpet(
-  client: Anthropic,
-  listingText: string,
-  listingUrl: string
-): Promise<{ hasTrumpet: boolean; summary: string | null }> {
-  const prompt = `You are checking a job listing to see if it specifically needs a trumpet player.
-
-Does this listing mention TRUMPET as an instrument sought? This includes:
-- Principal/Second/Section/Associate Trumpet
-- Trumpet substitute, extra, or sub
-- Brass auditions that explicitly include trumpet
-- General orchestral auditions where brass/trumpet players would audition
-
-Do NOT return true for: general musician listings with no instrument specification,
-listings only mentioning other brass (trombone, horn, tuba), or purely administrative roles.
-
-Return JSON only: { "hasTrumpet": boolean, "summary": string | null }
-- summary: 1-2 sentences describing the position (null if hasTrumpet is false)
-
-Listing URL: ${listingUrl}
-Content (first 4000 chars):
-${listingText.slice(0, 4000)}`;
-
-  try {
-    const response = await client.messages.create({
-      model: "claude-sonnet-4-20250514",
-      max_tokens: 400,
-      messages: [{ role: "user", content: prompt }],
-    });
-
-    const raw = response.content
-      .filter((b) => b.type === "text")
-      .map((b) => (b as { type: "text"; text: string }).text)
-      .join("");
-
-    const match = raw.match(/\{[\s\S]*\}/);
-    if (!match) throw new Error("No JSON object found");
-    return JSON.parse(match[0]) as { hasTrumpet: boolean; summary: string | null };
-  } catch (err) {
-    console.warn(`  ⚠️  checkListingForTrumpet failed: ${err}`);
-    return { hasTrumpet: false, summary: null };
-  }
-}
-
-// ─── Playbill: orchestrate two-level crawl ────────────────────────────────────
-
-/**
- * Coordinator for the Playbill multi-level crawl. Mutates `state` directly (passed
- * by reference) so that intermediate listing discoveries and trumpet verdicts are
- * captured even if the function is interrupted before returning — the caller saves
- * state to disk after this returns, but partial progress is preserved in the object.
- *
- * Two-tier cache strategy:
- *   1. Index page hash — if unchanged, skip re-extraction entirely (no Claude call).
- *      New listings can only appear if the index changes.
- *   2. `notified` flag per listing — drives which detail pages get fetched each run,
- *      regardless of the index hash. This handles the case where a prior run confirmed
- *      trumpet but the email send failed before `notified` was written.
- *
- * Non-trumpet listings are marked `notified=true` immediately to prevent redundant
- * detail-page fetches on future runs. Only trumpet-positive listings stay pending
- * until a successful email delivery is confirmed by the caller.
- */
-async function processPlaybillUrl(
-  client: Anthropic,
-  urlConfig: UrlConfig,
-  state: StateFile
-): Promise<PlaybillFinding[]> {
-  console.log(`\n📋 Playbill Job Board`);
-  const findings: PlaybillFinding[] = [];
-  const now = new Date().toISOString();
-
-  // 1. Fetch and hash the index page
-  let indexText: string;
-  try {
-    indexText = await scrapeUrl(urlConfig.url);
-  } catch (err) {
-    console.error(`  ❌ Failed to fetch Playbill index: ${err}`);
-    return findings;
-  }
-
-  const indexHash = contentHash(indexText);
-  const indexChanged = indexHash !== state.playbillIndexHash;
-
-  if (indexChanged) {
-    console.log(`  🔍 Index page changed — extracting musician listings...`);
-    state.playbillIndexHash = indexHash;
-
-    // 2. Extract musician listings from the index
-    const extracted = await extractPlaybillListings(client, indexText);
-    console.log(`  → Found ${extracted.length} musician listing(s) on index`);
-
-    // 3. Register any new listings in state (don't overwrite existing entries)
-    for (const item of extracted) {
-      if (!state.playbillListings[item.url]) {
-        console.log(`    ✨ New listing: "${item.title}" — ${item.organization}`);
-        state.playbillListings[item.url] = {
-          url: item.url,
-          title: item.title,
-          organization: item.organization,
-          firstSeen: now,
-          lastChecked: "",
-          hasTrumpet: false,
-          notified: false,
-          summary: null,
-        };
-      }
-    }
-  } else {
-    console.log(`  ✓ Index page unchanged — checking any pending listings`);
-  }
-
-  // 4. Process all unnotified listings (regardless of index change)
-  const pending = Object.values(state.playbillListings).filter((l) => !l.notified);
-  console.log(`  → ${pending.length} unnotified listing(s) to process`);
-
-  for (const listing of pending) {
-    // Already confirmed trumpet on a prior run but email failed — collect without re-fetching
-    if (listing.hasTrumpet) {
-      console.log(`    🎺 "${listing.title}" — trumpet confirmed (pending notification)`);
-      findings.push({
-        listingUrl: listing.url,
-        title: listing.title,
-        organization: listing.organization,
-        summary: listing.summary,
-      });
-      continue;
-    }
-
-    // Fetch detail page and check for trumpet
-    console.log(`    Checking: "${listing.title}"`);
-    try {
-      const detailText = await scrapeUrl(listing.url);
-      const { hasTrumpet, summary } = await checkListingForTrumpet(client, detailText, listing.url);
-
-      listing.lastChecked = now;
-      listing.hasTrumpet = hasTrumpet;
-      listing.summary = summary;
-
-      if (hasTrumpet) {
-        console.log(`    🎺 Trumpet found! "${listing.title}"`);
-        findings.push({
-          listingUrl: listing.url,
-          title: listing.title,
-          organization: listing.organization,
-          summary,
-        });
-      } else {
-        console.log(`    ✗ No trumpet mention`);
-        // Mark notified=true to avoid rechecking non-trumpet listings every run
-        listing.notified = true;
-      }
-    } catch (err) {
-      console.warn(`    ⚠️  Could not process listing "${listing.title}": ${err}`);
-    }
-  }
-
-  return findings;
-}
-
 // ─── State helpers ────────────────────────────────────────────────────────────
 
 /**
@@ -574,10 +194,6 @@ function loadState(): StateFile {
 
 function saveState(state: StateFile): void {
   fs.writeFileSync(STATE_FILE, JSON.stringify(state, null, 2), "utf-8");
-}
-
-function contentHash(text: string): string {
-  return crypto.createHash("sha256").update(text).digest("hex").slice(0, 16);
 }
 
 // ─── Email (Gmail API + OAuth2) ───────────────────────────────────────────────
@@ -956,7 +572,6 @@ async function preflightUrls(
     failures,
   };
 }
-
 
 // ─── GitHub Issue ─────────────────────────────────────────────────────────────
